refactor(pre): preliminary refactor
