+++ src/routes/transfer/transfer.js
@@ -1,6 +1,7 @@
 //@ts-check
-
-import { transferUSDC } from '../../utils/transfer-spl.js';
+import BigNumber from 'bignumber.js';
+import { createUserTokenAccount, getUserTokenAccount, transferUSDC } from '../../repos/transfer.js';
+import { MOCHA_KEYPAIR, deriveAddress } from '../../utils/solana.js';
 
 /**
  * @param {string} fromNumber
@@ -14,6 +15,36 @@ export async function transfer(fromNumber, toNumber, amount) {
   // https://twilio.com/docs/lookup/quickstart
   //https://www.twilio.com/docs/glossary/what-e164
   console.log(`Transferrring USDC`, amount);
+
+  const fromWhatsappUserAccount = await getOrCreateUserTokenAccount(fromNumber);
+  const toWhatsappUserAccount = await getOrCreateUserTokenAccount(toNumber);
+
+  // fix amount
+  let parsedAmount = new BigNumber(amount).multipliedBy(10 ** 6).toNumber();
+
   // call transfer
-  return transferUSDC(fromNumber, toNumber, amount);
+  return transferUSDC(MOCHA_KEYPAIR, fromWhatsappUserAccount, toWhatsappUserAccount, parsedAmount);
+}
+
+/**
+ * @description Function to fetch the wallet balance
+ * @param {string} phoneNumber
+ */
+async function getOrCreateUserTokenAccount(phoneNumber) {
+  const address = await deriveAddress(MOCHA_KEYPAIR.publicKey, phoneNumber);
+
+  try {
+    const account = await getUserTokenAccount(address);
+    console.log('Account found', account.address.toBase58());
+
+    return account.address;
+  } catch (error) {
+    console.log('Account not found, creating account ...', {
+      phoneNumber,
+      address: address.toBase58(),
+    });
+
+    const newAccount = await createUserTokenAccount(MOCHA_KEYPAIR, address, phoneNumber);
+    return newAccount.address;
+  }
 }
